Rebase
added docsting to 2 new functions
